requirements.txt
aioconsole == 0.7.1
arrow == 1.3.0
colorama == 0.4.6
# Branch made to support directly using escape codes instead of just predefined color names
colorlog @ git+https://github.com/svioletg/python-colorlog@main
discord.py == 2.4.0
discord-pretty-help == 2.0.7
fuzzywuzzy == 0.18.0
pylast == 5.3.0
PyNaCl == 1.5.0
python-benedict == 0.33.2
python-Levenshtein == 0.25.1
pytube == 15.0.0
PyYAML == 6.0.1
soundcloud_lib == 0.6.1
spotipy == 2.24.0
requests == 2.32.2
yt_dlp == 2024.7.2
ytmusicapi == 1.7.4